// src/sunnylink/v1/schema_athena.d.ts
export interface components {
	schemas: {
// ...
		DeviceParam: {
			key?: string;
			is_compressed?: boolean;
			value?: string;
			type?: components['schemas']['ParamType'];
		};
// ...
		ParamType: 'String' | 'Bool' | 'Int' | 'Float' | 'Time' | 'Json' | 'Bytes' | 'Unknown';
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}